Working on The Frontend - testing the final code
To test, put "yarn start" in the powershell in the file of the code (from GitHub desktop, go to repository, show in explorer). Then in the file explorer, click file, open powershell, run as administrator, and type in "yarn start." Make sure the Sunrise and Sunset application in IntelliJ is running. If it doesn't work, the problem might be in the "displaying the data within the search result" in the instructions.